requirements.txt
# Core LlamaIndex and OpenAI dependencies
llama-index>=0.10.20
llama-index-core>=0.10.20
llama-index-llms-openai>=0.1.12
llama-index-readers-web>=0.1.6
llama-index-embeddings-openai>=0.1.11
openai>=1.12.0

# Web scraping and HTTP
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.1

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Environment and utilities
python-dotenv==1.0.1

# Optional: For enterprise features (uncomment if needed)
# redis==5.0.1
# pyjwt==2.8.0 